jax/requirements_kaggle.txt
# JAX and Flax for TPU
# Use compatible versions for Kaggle TPU
# Note: Kaggle may have pre-installed JAX, check version first
jax[tpu]>=0.4.20,<0.4.30
flax>=0.7.5,<0.9.0
optax>=0.1.7,<0.2.0

# Diffusers for VAE
diffusers==0.25.1
transformers>=4.35.0

# TensorFlow for data pipeline and FID
tensorflow==2.15.0
tensorflow-hub==0.15.0

# Image processing
Pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0

# Optional: Better array operations
einops>=0.7.0

# Logging and experiment tracking
wandb>=0.16.0

# YAML config
PyYAML>=6.0

# Progress bars
tqdm>=4.66.0

# Type checking (optional, for development)
# jaxtyping>=0.2.0
# typeguard>=4.0.0
